+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added the `Tab` component.
 - Added the `ToggleButtonGroup` component.
 - Added the `ProgressBar` component.
+- Added `selected`, `showEndIcon`, `label` properties in `IconButton` component
 
 ### Fixed
 
@@ -29,6 +30,7 @@
 - Adjusted the text alignment of the Tab component.
 - Added hover actions for the Accordion component.
 - Added sync icon for the Tile component.
+- Adjusted `IconButton` component styling
 
 ### Breaking changes